requirements.txt
matplotlib==3.1.3
numpy==1.18.2
Pillow==9.0.1
pydicom==1.2.0
torch==1.11.0
torchvision==0.12.0
